Renaming and Adding Notes to Jupyter Notebook
Was side tracked from supposed purpose of this branch, i.e., to modularize the code for ease of tracking changes, as well as creating new features, e.g., making more VQE runs to test for convergence. This change is supposed to go to a different branch, i.e., making a documented version of my thesis codes for replicability. Will create a separate branch for this later on.