Generated Position Basis Data
Wrote a short python script that's supposed to check for convergence and ran this to generate data sets for all position basis Hamiltonians across each ansatz. Will choose data from this to be graphed as part of the final manuscript. Next step is to do the same for the energy basis Hamiltonians.