Edited core module + flowchart + crucial bug
Created a flowchart to help give ideas how to write the manuscript flowchart. Also streamlined code, i.e. removed paramnums in runVQE function as it's distracting and is already given by the type of ansatz. Found a crucial bug where only n points are generated instead of n + 1 points, where n is dimensionality of the parameter space. This was taken from the main reference. Changed this, but convergence was already shown and not sure how this changes anything. Will have to rerun entire code, though.